+++ Libraries/Bootstrap.php
@@ -14,26 +14,25 @@ namespace c975L\IncludeLibraryBundle\Libraries;
  */
 class Bootstrap
 {
-    public function getData($type, $version)
-    {
-        //Gets css
-        if ($type == 'css') {
-            return $this->getCss($version);
-        //Gets javascript
-        } elseif ($type == 'javascript') {
-            return $this->getJavascript($version);
-        }
-
-        return;
-    }
-
     /*
-     * Refer to README.md for how-to add versions/library
+     * Refer to README.md for how-to add versions
      */
     public function getCss($version)
     {
         switch ($version) {
             case 'latest':
+
+            case '4.*':
+            case '4.0.*':
+            case '4.0.0':
+            case '4.0.0.*':
+                $data = array(
+                    'href' => 'https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css',
+                    'integrity' => 'sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm',
+                    'crossorigin' => 'anonymous',
+                );
+                break;
+
             case '3.*':
             case '3.3.*':
             case '3.3.7':
@@ -44,8 +43,9 @@ class Bootstrap
                     'crossorigin' => 'anonymous',
                 );
                 break;
+
             default:
-                $data = '';
+                $data = null;
                 break;
         }
 
@@ -53,12 +53,24 @@ class Bootstrap
     }
 
     /*
-     * Refer to README.md for how-to add versions/library
+     * Refer to README.md for how-to add versions
      */
     public function getJavascript($version)
     {
         switch ($version) {
             case 'latest':
+
+            case '4.*':
+            case '4.0.*':
+            case '4.0.0':
+            case '4.0.0.*':
+                $data = array(
+                    'src' => 'https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js',
+                    'integrity' => 'sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl',
+                    'crossorigin' => 'anonymous',
+                );
+                break;
+
             case '3.*':
             case '3.3.*':
             case '3.3.7':
@@ -69,8 +81,9 @@ class Bootstrap
                     'crossorigin' => 'anonymous',
                 );
                 break;
+
             default:
-                $data = '';
+                $data = null;
                 break;
         }