+++ Libraries/Popper.php
@@ -1,0 +1,43 @@
+<?php
+/*
+ * (c) 2018: 975L <contact@975l.com>
+ * (c) 2018: Laurent Marquet <laurent.marquet@laposte.net>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace c975L\IncludeLibraryBundle\Libraries;
+
+/*
+ * jQuery - https://popper.js.org
+ */
+class Popper
+{
+    /*
+     * Refer to README.md for how-to add versions
+     */
+    public function getJavascript($version)
+    {
+        switch ($version) {
+            case 'latest':
+
+            case '1.*':
+            case '1.12.*':
+            case '1.12.9':
+            case '1.12.9.*':
+                $data = array(
+                    'src' => 'https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.min.js',
+                    'integrity' => 'sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q',
+                    'crossorigin' => 'anonymous',
+                );
+                break;
+
+            default:
+                $data = null;
+                break;
+        }
+
+        return $data;
+    }
+}