+++ ChangeLog.md
@@ -1,5 +1,20 @@
 # Changelog
 
+v1.3
+----
+- Added corrections and re-formatted `README.md` (24/02/2018)
+- Added Bootstrap 4.0.0 (24/02/2018)
+- Added Popper 1.12.9 library (24/02/2018)
+- Corrected test to check existence of data in Twig templates (24/02/2018)
+- Throws an error if a library is not found (24/02/2018)
+- Corrected version of Cookieconsent + web site (24/02/2018)
+- Added defer attribute for javascript files (24/02/2018)
+- Added javascript for Fontawesome (SVG) (24/02/2018)
+- Removed `getData()` method in libraries classes and replaced by a direct call in service (24/02/2018)
+- Added jQuery Slim (24/02/2018)
+- Added Fontawesome brands + solids (24/02/2018)
+- Added jQueryUI 1.12.1 (24/02/2018)
+
 v1.2
 ----
 - Added wildcards values for existing libraries (23/02/2018)